merge new version

+++ src/sentry/web/frontend/accounts_twofactor.py
@@ -1,3 +1,5 @@
+from __future__ import absolute_import
+
 from django import forms
 from django.db import transaction
 from django.http import HttpResponseRedirect, Http404
@@ -8,19 +10,27 @@ from django.utils.decorators import method_decorator
 from django.core.context_processors import csrf
 from django.utils.translation import ugettext_lazy as _
 
+import petname
 from sudo.decorators import sudo_required
 
 from sentry.models import Authenticator
 from sentry.web.frontend.base import BaseView
 from sentry.web.decorators import login_required
 from sentry.web.helpers import render_to_response
-from sentry.web.forms.accounts import TwoFactorForm
+from sentry.web.forms.accounts import TwoFactorForm, ConfirmPasswordForm
 from sentry.utils import json
 
 
 class SmsForm(forms.Form):
     phone_number = forms.CharField(
         label=_('Phone number'), max_length=40
+    )
+
+
+class U2fForm(forms.Form):
+    device_name = forms.CharField(
+        label=_('Device name'), max_length=60, required=False,
+        initial=lambda: petname.Generate(2, ' ').title(),
     )
 
 
@@ -38,14 +48,12 @@ class TwoFactorSettingsView(BaseView):
                 request.user, self.interface_id)
         except LookupError:
             raise Http404
-        if 'remove' in request.POST:
-            return self.remove(request, interface)
         return self.configure(request, interface)
 
     def make_context(self, request, interface):
         context = csrf(request)
         context['auth'] = interface
-        context['page'] = 'settings'
+        context['page'] = 'security'
         return context
 
     def delete_authenticator(self, interface):
@@ -70,12 +78,23 @@ class TwoFactorSettingsView(BaseView):
                 iface.authenticator.delete()
 
     def remove(self, request, interface):
+        form = ConfirmPasswordForm(request.user)
+
         if 'no' in request.POST or \
            not interface.is_enrolled:
             return HttpResponseRedirect(reverse('sentry-account-settings-2fa'))
         elif 'yes' in request.POST:
-            self.delete_authenticator(interface)
-            return HttpResponseRedirect(reverse('sentry-account-settings-2fa'))
+            form = ConfirmPasswordForm(request.user, request.POST)
+            if 'password' in form.fields:
+                if form.is_valid():
+                    if request.user.check_password(form.cleaned_data['password']):
+                        self.delete_authenticator(interface)
+                        return HttpResponseRedirect(reverse('sentry-account-settings-2fa'))
+                    else:
+                        form.errors['__all__'] = ['Invalid password.']
+            else:
+                self.delete_authenticator(interface)
+                return HttpResponseRedirect(reverse('sentry-account-settings-2fa'))
 
         all_interfaces = Authenticator.objects.all_interfaces_for_user(
             request.user)
@@ -86,6 +105,7 @@ class TwoFactorSettingsView(BaseView):
             len(backup_interfaces) == len(other_interfaces)
 
         context = self.make_context(request, interface)
+        context['password_form'] = form
         context['removes_backups'] = removes_backups
         return render_to_response('sentry/account/twofactor/remove.html',
                                   context, request)
@@ -97,12 +117,23 @@ class TwoFactorSettingsView(BaseView):
         if not insecure \
            or (interface.is_backup_interface and
                Authenticator.objects.user_has_2fa(request.user)):
-            interface.enroll(request.user)
-            if Authenticator.objects.auto_add_recovery_codes(request.user):
-                next = reverse('sentry-account-settings-2fa-recovery')
+            try:
+                interface.enroll(request.user)
+            except Authenticator.AlreadyEnrolled:
+                # This can happen in some cases when races occur.  We have
+                # seen this when people press the submit button twice.  In
+                # that case just go to the overview page of 2fa
+                next = reverse('sentry-account-settings-2fa')
+            else:
+                request.user.refresh_session_nonce(self.request)
+                request.user.save()
+                if Authenticator.objects.auto_add_recovery_codes(request.user):
+                    next = reverse('sentry-account-settings-2fa-recovery')
         return HttpResponseRedirect(next)
 
     def configure(self, request, interface):
+        if 'remove' in request.POST:
+            return self.remove(request, interface)
         if 'enroll' in request.POST or \
            request.GET.get('enroll') == 'yes':
             return self.enroll(request, interface,
@@ -134,16 +165,30 @@ class TotpSettingsView(TwoFactorSettingsView):
 
         if 'otp' in request.POST:
             form = TwoFactorForm(request.POST)
-            if form.is_valid() and interface.validate_otp(
-                    form.cleaned_data['otp']):
-                return TwoFactorSettingsView.enroll(self, request, interface)
+            password_form = ConfirmPasswordForm(request.user, request.POST)
+            if 'password' in password_form.fields and password_form.is_valid():
+                if request.user.check_password(password_form.cleaned_data['password']):
+                    if form.is_valid() and interface.validate_otp(
+                            form.cleaned_data['otp']):
+                        return TwoFactorSettingsView.enroll(self, request, interface)
+                    else:
+                        form.errors['__all__'] = ['Invalid confirmation code.']
+                else:
+                    form.errors['__all__'] = ['Invalid password.']
             else:
-                form.errors['__all__'] = ['Invalid confirmation code.']
+                if form.is_valid() and interface.validate_otp(
+                        form.cleaned_data['otp']):
+                    return TwoFactorSettingsView.enroll(self, request, interface)
+                else:
+                    form.errors['__all__'] = ['Invalid confirmation code.']
+
         else:
             form = TwoFactorForm()
+            password_form = ConfirmPasswordForm(request.user)
 
         context = self.make_context(request, interface)
         context['otp_form'] = form
+        context['password_form'] = password_form
         context['provision_qrcode'] = interface.get_provision_qrcode(
             request.user.email)
         return render_to_response('sentry/account/twofactor/enroll_totp.html',
@@ -191,16 +236,39 @@ class SmsSettingsView(TwoFactorSettingsView):
 class U2fSettingsView(TwoFactorSettingsView):
     interface_id = 'u2f'
 
+    def configure(self, request, interface):
+        # Try to remove a key handle.  If this returns `False` it means we
+        # are about to remove the last key handle.  In that case just
+        # bubble through to the configure page which will pick up the
+        # 'remove' in the form and bring up the remove screen for the
+        # entire authentication method.
+        key_handle = request.POST.get('key_handle')
+        if key_handle and 'remove' in request.POST and \
+           interface.remove_u2f_device(key_handle):
+            interface.authenticator.save()
+            return HttpResponseRedirect(request.path)
+
+        return TwoFactorSettingsView.configure(self, request, interface)
+
     def enroll(self, request, interface, insecure=False):
+        u2f_form = U2fForm()
+
         challenge = request.POST.get('challenge')
         if challenge:
-            interface.enrollment_data = json.loads(challenge)
+            enrollment_data = json.loads(challenge)
+        else:
+            enrollment_data = interface.start_enrollment()
 
         response = request.POST.get('response')
         if response:
-            interface.try_enroll(json.loads(response))
-            return TwoFactorSettingsView.enroll(self, request, interface)
+            u2f_form = U2fForm(request.POST)
+            if u2f_form.is_valid():
+                interface.try_enroll(enrollment_data, json.loads(response),
+                                     u2f_form.cleaned_data['device_name'])
+                return TwoFactorSettingsView.enroll(self, request, interface)
 
         context = self.make_context(request, interface)
+        context['enrollment_data'] = enrollment_data
+        context['u2f_form'] = u2f_form
         return render_to_response('sentry/account/twofactor/enroll_u2f.html',
                                   context, request)
